【新增】分配前台逻辑【songkaojun】

+++ src/main/java/cn/explink/enumutil/express/ExcuteStateEnum.java
@@ -15,7 +15,7 @@ public enum ExcuteStateEnum {
 		this.text = text;
 	}
 
-	public long getValue() {
+	public int getValue() {
 		return this.value;
 	}
